pyproject.toml
[project]
name = "newspapergenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langsmith>=0.4.31",
    "openai>=2.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.9",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "sentence-transformers>=5.1.1",
    "tiktoken>=0.11.0",
]
